Initial MCST implementation
Implemented the Monte Carlo Search Tree as descibed in the AlphaGo Zero paper, and added a cell to the notebook to run it. The neural network is not implemented, so that will need to be added in once it has been. A config file was also added to keep track of magic numbers. The paper also descriibes adding noise when selecting a move, which has not been implemented yet.